Renamed utils/logo to utils/branding. Consolidated utils/misc into utils/console and renamed it to utils/logging.
