+++ src/bin/aunty.js
@@ -5,8 +5,8 @@ const resolve = require('resolve');
 
 // Ours
 const {name} = require('../../package');
-const {error} = require('../utils/console');
-const {createErrorLogo} = require('../utils/logo');
+const {createErrorLogo} = require('../utils/branding');
+const {error} = require('../utils/logging');
 
 function exit(err) {
   if (err) {
