+++ src/commands/build-basic-story/index.js
@@ -11,7 +11,7 @@ const vfs = require('vinyl-fs');
 const {command} = require('../../cli');
 const {pumped, throws} = require('../../utils/async');
 const {bad, cmd, ok} = require('../../utils/color');
-const {log, warn} = require('../../utils/console');
+const {log, warn} = require('../../utils/logging');
 const {styleLastSegment} = require('../../utils/strings');
 const {clean} = require('../clean');
 const {OPTIONS, KEY, D_KEY, DEFAULTS, MESSAGES} = require('./constants');
