+++ src/commands/new/constants.js
@@ -5,7 +5,7 @@ const {relative} = require('path');
 const ourPkg = require('../../../package');
 const {BUILD_DIR, PROJECT_TYPES, PROJECT_TYPE_DESCRIPTIONS} = require('../../projects/constants');
 const {cmd, hvy, ok, opt, req, sec} = require('../../utils/color');
-const {pretty} = require('../../utils/misc');
+const {pretty} = require('../../utils/logging');
 const {indented, listPairs} = require('../../utils/strings');
 
 const [MAJOR, MINOR] = ourPkg.version.split('.');
